requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.6.4
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.9
duckdb==0.8.1
exceptiongroup==1.1.1
fastapi==0.99.0
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.35.2
gradio_client==0.2.7
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.220
langchainplus-sdk==0.0.19
linkify-it-py==2.0.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.1
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
Pillow==9.5.0
posthog==3.0.1
pycparser==2.21
pydantic==1.10.10
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
